pyproject.toml
[tool.poetry]
name = "orc-agent"
version = "1.0.0"
description = "Automated documentation generation system for GitHub repositories"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
redis = {extras = ["hiredis"], version = "^5.0.1"}
aiohttp = "^3.9.1"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
openai = "^1.10.0"
pygithub = "^2.1.1"
prometheus-client = "^0.19.0"
jinja2 = "^3.1.3"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
httpx = "^0.26.0"
langchain = "^0.1.0"
langchain-community = "^0.0.38"
langchain-openai = "^0.0.5"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
black = "^24.1.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.8.0"
types-redis = "^4.6.0"
aioresponses = "^0.7.6"
faker = "^22.5.0"

[tool.poetry.scripts]
start = "uvicorn src.main:app --host 0.0.0.0 --port 8000"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "fastapi.*",
    "uvicorn.*",
    "redis.*",
    "prometheus_client.*",
    "pytest.*",
    "langchain.*",
    "langchain_community.*",
    "langchain_openai.*"
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=src --cov-report=xml --cov-report=term-missing"
markers = [
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests"
] 